Implemented fixed-size block matrices.
Unfortunately, LIS currently only supports user-defined arrays for CSR
matrices, so we might have to either help them out, or find another way
to navigate block rows/columns. I'm inclined to go with the former, and
contribute BSR and VBR implementations of the global-to-local index mappings
for them, but this requires some assessment.