Numerous subtleties addressed in PETSc/HYPRE solvers.
Getting these packages to compute and use residual norms consistently is
a challenge that I hadn't heretofore appreciated. But now they can solve
our Newton test problems using the INK method. LIS is still a work in
progress, but we are inching closer with a patch to lis_solver for fixing
its treatment of residual norms in convergence.