I/O tests are demanding a feature I wanted to defer.
Since the planar_polymesh type is not distributed, it's necessary to allow the
specification of a global (planar) domain by a union of distributed planar
polymesh objects. The global domain must be constructed without creating a
global representation of the union of these planar polymeshes. This requires
create_empty_colmesh_from_fragments().